Change newrelic-log-ingestion install mechanism
Instead of a stack the refers to a SAR app, just install
the SAR app. This removes the need for CAPABILITY_AUTO_EXPAND.

Because of this change, we also reworked the upgrade code. It can
now unwrap a nested install, which makes it look like the new-style
install. The CLI is a little different now as well: you don't have
to re-specify install options during an upgrade. The default is to
preserve your current values.